% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names2sex.R
\name{full2first}
\alias{full2first}
\title{Convert full name string to first name}
\usage{
full2first(x)
}
\arguments{
\item{x}{A character vector of full names (each full name is a string)}
}
\value{
A vector of first names
}
\description{
Parses full names vector into first names vector
}
