% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name2sex-package.R
\docType{package}
\name{name2sex-package}
\alias{name2sex}
\alias{name2sex-package}
\title{name2sex: Convert Names to Sex}
\description{
Takes first names and looks up the percent male and female using
public data bases.
}
\author{
\strong{Maintainer}: Michael Wayne Kearney \email{kearneymw@missouri.edu} (0000-0002-0730-4694)

Authors:
\itemize{
  \item Lingshu Hu \email{lhn95@mail.missouri.edu}
}

}
\keyword{internal}
