% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names2sex.R
\name{names2sex}
\alias{names2sex}
\title{Convert names to sex estimates}
\usage{
names2sex(nm)
}
\arguments{
\item{nm}{Vector of first names}
}
\value{
A data frame with names, percent female, and sex variables
}
\description{
Takes first names (input) and looks up percent female/male using national
data from the Social Security Administration
}
